+++ integrations/oci/metrics/cdi/src/main/java/io/helidon/integrations/oci/metrics/cdi/OciMetricsBean.java
@@ -17,6 +17,7 @@ package io.helidon.integrations.oci.metrics.cdi;
 
 import io.helidon.config.Config;
 import io.helidon.integrations.oci.metrics.OciMetricsSupport;
+import io.helidon.integrations.oci.metrics.OciMetricsSupportFactory;
 import io.helidon.microprofile.server.RoutingBuilders;
 
 import com.oracle.bmc.monitoring.Monitoring;
@@ -38,9 +39,8 @@ import static jakarta.interceptor.Interceptor.Priority.LIBRARY_BEFORE;
 
 // This bean is added to handle injection on the ObserverMethod as it does not work on an Extension class.
 @Singleton
-public class OciMetricsBean {
+public class OciMetricsBean extends OciMetricsSupportFactory {
 
-    private Config ociMetricsConfig;
     private OciMetricsSupport ociMetricsSupport;
 
     /**
@@ -50,63 +50,23 @@ public class OciMetricsBean {
     public OciMetricsBean() {
     }
 
-    /**
-     * Returns the config key to use for retrieving OCI metrics settings from the root config.
-     *
-     * @return config key for OCI metrics settings
-     */
+    @Override
     protected String configKey() {
         return "ocimetrics";
     }
 
-    /**
-     * Returns the builder for constructing a new {@link io.helidon.integrations.oci.metrics.OciMetricsSupport} instance,
-     * initialized using the config retrieved using the {@link #configKey()} return value and the provided
-     * {@link com.oracle.bmc.monitoring.Monitoring} instance.
-     *
-     * @param rootConfig root {@link io.helidon.config.Config} node
-     * @param ociMetricsConfig config node for the OCI metrics settings
-     * @param monitoring monitoring implementation to be used in preparing the {@code OciMetricsSupport.Builder}.
-     * @return resulting builder
-     */
-    protected OciMetricsSupport.Builder ociMetricsSupportBuilder(Config rootConfig,
-                                                                 Config ociMetricsConfig,
-                                                                 Monitoring monitoring) {
-        return OciMetricsSupport.builder()
-                .config(ociMetricsConfig)
-                .monitoringClient(monitoring);
-    }
-
-    /**
-     * Returns the OCI metrics config settings previously retrieved from the config root.
-     *
-     * @return  OCI metrics config node
-     */
-    protected Config ociMetricsConfig() {
-        return ociMetricsConfig;
-    }
-
-    /**
-     * Activates OCI metrics support.
-     *
-     * @param rootConfig root config node
-     * @param ociMetricsConfig OCI metrics configuration
-     * @param builder builder for {@link io.helidon.integrations.oci.metrics.OciMetricsSupport}
-     */
+    @Override
     protected void activateOciMetricsSupport(Config rootConfig, Config ociMetricsConfig, OciMetricsSupport.Builder builder) {
-        ociMetricsSupport = builder.build();
+        this.ociMetricsSupport = builder.build();
         RoutingBuilders.create(ociMetricsConfig)
                 .routingBuilder()
                 .register(ociMetricsSupport);
     }
+
     // Make Priority higher than MetricsCdiExtension so this will only start after MetricsCdiExtension has completed.
     void registerOciMetrics(@Observes @Priority(LIBRARY_BEFORE + 20) @Initialized(ApplicationScoped.class) Object ignore,
                             Config rootConfig, Monitoring monitoringClient) {
-        ociMetricsConfig = rootConfig.get(configKey());
-        OciMetricsSupport.Builder builder = ociMetricsSupportBuilder(rootConfig, ociMetricsConfig, monitoringClient);
-        if (builder.enabled()) {
-            activateOciMetricsSupport(rootConfig, ociMetricsConfig, builder);
-        }
+        registerOciMetrics(rootConfig, monitoringClient);
     }
 
     // For testing
