+++ integrations/oci/metrics/cdi/src/main/java/io/helidon/integrations/oci/metrics/cdi/OciMetricsCdiExtension.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2022, 2023 Oracle and/or its affiliates.
+ * Copyright (c) 2022, 2024 Oracle and/or its affiliates.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -23,6 +23,13 @@ import jakarta.enterprise.inject.spi.Extension;
  * OCI metrics integration CDI extension.
  */
 public class OciMetricsCdiExtension implements Extension {
+
+    /**
+     * For CDI use only.
+     */
+    @Deprecated
+    public OciMetricsCdiExtension() {
+    }
 
     // A new bean is added to handle the Observer Method as injection does not work here
     void addOciMetricsBean(@Observes BeforeBeanDiscovery event) {
