+++ integrations/oci/metrics/metrics/src/main/java/io/helidon/integrations/oci/metrics/OciMetricsSupport.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2022, 2023 Oracle and/or its affiliates.
+ * Copyright (c) 2022, 2024 Oracle and/or its affiliates.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -302,6 +302,9 @@ public class OciMetricsSupport implements HttpService {
         private int batchSize = DEFAULT_BATCH_SIZE;
         private boolean enabled = true;
         private Monitoring monitoringClient;
+
+        private Builder() {
+        }
 
         @Override
         public OciMetricsSupport build() {
